Update zettelID implementation
zettelID was also embedded into the title of the post.
With this change, it's now decoupled and can be used separately.
As an example, metadata used by search engine crawlers don't need
zettelID in the post. Now post titles don't include zettelID
on the metadata while keeping it in the post's title.